refactor: decouple package outputs
